Add final polish

+++ src/heap.h
@@ -4,17 +4,17 @@
 
 typedef struct {
     Node *node;
-    double score;
+    float score;
 } Element;
 
 typedef struct {
-    size_t capacity;
-    size_t n_elements;
+    unsigned capacity;
+    unsigned n_elements;
     Element *elements;
 } Heap;
 
 Heap *heap_init();
-void heap_push(Heap *const heap, Node *const node, const double score);
+void heap_push(Heap *const heap, Node *const node, const float score);
 Node *heap_pop(Heap *const heap);
-void heap_replace(Heap *const heap, Node *const node, const double score);
+void heap_replace(Heap *const heap, Node *const node, const float score);
 Bool heap_is_empty(const Heap *const heap);
